% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_ep.R
\name{special_ep}
\alias{special_ep}
\title{Find out if an episode of RuPaul's Drag Race was a special episode}
\usage{
special_ep(character, numeric)
}
\arguments{
\item{character}{The number of the season}

\item{numeric}{The number of the episode}
}
\value{
A character vector
}
\description{
Find out if an episode of RuPaul's Drag Race was a special episode
}
\examples{
library(tidyverse)
library(dragracer)

special_ep("S07", 9)

"It's not a special episode"


special_ep("S14", 11)

"It's a special episode hunty!"

}
