% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winner.R
\name{winner}
\alias{winner}
\title{Find out which queen won a season}
\usage{
winner(season)
}
\arguments{
\item{season}{The number of the season}
}
\value{
A character vector
}
\description{
Find out which queen won a season
}
\examples{
library(tidyverse)
library(dragracer)

winner("S06")

"The queen who won this season was Bianca Del Rio"

}
