% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_info.R
\name{general_info}
\alias{general_info}
\title{Gain some general information about contestants on RuPaul's Drag Race}
\usage{
general_info(character)
}
\arguments{
\item{character}{A name of a contestant of RuPaul's Drag Race}
}
\value{
A character vector
}
\description{
Gain some general information about contestants on RuPaul's Drag Race
}
\examples{
library(tidyverse)
library(dragracer)

general_info("Jinkx Monsoon")
}
